Move security-controller.js from ~/interfaces to ~/application
