+++ api/lib/application/assessment-results/index.js
@@ -1,5 +1,5 @@
 const AssessmentResultController = require('./assessment-result-controller');
-const securityController = require('../../interfaces/controllers/security-controller');
+const securityController = require('../security-controller');
 
 exports.register = async function(server) {
   server.route([
