+++ api/lib/application/cache/index.js
@@ -1,5 +1,5 @@
 const Joi = require('@hapi/joi');
-const securityController = require('../../interfaces/controllers/security-controller');
+const securityController = require('../security-controller');
 const CacheController = require('./cache-controller');
 
 exports.register = async function(server) {
