+++ api/lib/application/certification-center-memberships/index.js
@@ -1,5 +1,5 @@
 const certificationCenterMembershipController = require('./certification-center-membership-controller');
-const securityController = require('../../interfaces/controllers/security-controller');
+const securityController = require('../security-controller');
 
 exports.register = async function(server) {
   server.route([
