+++ api/lib/application/certification-centers/index.js
@@ -1,5 +1,5 @@
 const certificationCenterController = require('./certification-center-controller');
-const securityController = require('../../interfaces/controllers/security-controller');
+const securityController = require('../security-controller');
 
 exports.register = async function(server) {
   server.route([
