+++ api/lib/application/certification-courses/index.js
@@ -1,4 +1,4 @@
-const securityController = require('../../interfaces/controllers/security-controller');
+const securityController = require('../security-controller');
 const certificationCourseController = require('./certification-course-controller');
 
 exports.register = async function(server) {
