+++ api/lib/application/memberships/index.js
@@ -1,4 +1,4 @@
-const securityController = require('../../interfaces/controllers/security-controller');
+const securityController = require('../security-controller');
 const membershipController = require('./membership-controller');
 
 exports.register = async function(server) {
