+++ api/tests/unit/application/organizations/index_test.js
@@ -1,6 +1,6 @@
 const { expect, sinon, HttpTestServer } = require('../../../test-helper');
 
-const securityController = require('../../../../lib/interfaces/controllers/security-controller');
+const securityController = require('../../../../lib/application/security-controller');
 const organizationController = require('../../../../lib/application/organizations/organization-controller');
 const usecases = require ('../../../../lib/domain/usecases');
