+++ api/tests/unit/application/user-tutorials/index_test.js
@@ -1,6 +1,6 @@
 const { expect, sinon } = require('../../../test-helper');
 const Hapi = require('@hapi/hapi');
-const securityController = require('../../../../lib/interfaces/controllers/security-controller');
+const securityController = require('../../../../lib/application/security-controller');
 const userTutorialsController = require('../../../../lib/application/user-tutorials/user-tutorials-controller');
 
 let server;
